numeric and promotion
