+++ aten/src/ATen/native/special_functions/detail/tan_pi.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ATen/native/special_functions/detail/numeric_t.h>
+#include <ATen/native/special_functions/detail/numeric.h>
 #include <c10/util/MathConstants.h>
 
 namespace at {
