+++ aten/src/ATen/native/special_functions/sinhc.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ATen/native/special_functions/detail/promote_t.h>
+#include <ATen/native/special_functions/detail/promote.h>
 #include <ATen/native/special_functions/detail/sinhc.h>
 #include <c10/macros/Macros.h>
