+++ aten/src/ATen/native/special_functions/tan_pi.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ATen/native/special_functions/detail/promote_t.h>
+#include <ATen/native/special_functions/detail/promote.h>
 #include <ATen/native/special_functions/detail/tan_pi.h>
 #include <c10/macros/Macros.h>
 #include <c10/util/complex.h>
